worker: no changes, fixes diff
